+++ R/contrast.R
@@ -382,7 +382,7 @@ contrast.emmGrid = function(object, method = "eff", interaction = FALSE,
     }
     misc$methDesc = attr(cmat, "desc")
     misc$famSize = ifelse(is.null(fs <- attr(cmat, "famSize")), length(by.rows[[1]]), fs)
-    misc$pri.vars = setdiff(names(grid), c(".offset.",".wgt."))
+    misc$pri.vars = setdiff(names(grid), c(by, ".offset.",".wgt."))
     if (missing(adjust)) adjust = attr(cmat, "adjust")
     if (is.null(adjust)) adjust = "none"
     if (!is.null(attr(cmat, "offset")))
@@ -398,6 +398,8 @@ contrast.emmGrid = function(object, method = "eff", interaction = FALSE,
     misc$adjust = adjust
     misc$infer = c(FALSE, TRUE)
     misc$by.vars = by
+    if(misc$estType == "pairs") # internal flag to keep track of original by vars for paired comps
+        misc$.pairby = paste(c("", by), collapse = ",")
     # save contrast coefs
     by.cols = seq_len(ncol(tcmat))
     if(!is.null(by.rows))