+++ R/transformations.R
@@ -364,3 +364,27 @@ make.tran = function(type = c("genlog", "power", "boxcox", "sympower",
     result
 }
 
+
+### Internal routine to make a scales::trans object
+.make.scale = function(misc) {
+    if (!requireNamespace("scales", quiet = TRUE)) 
+        stop("type = \"scale\" requires the 'scales' package to be installed")
+    tran = misc$tran
+    if (is.character(tran)) {
+        # is it a canned scale?
+        if ((length(intersect(names(misc), c("tran.mult", "tran.offset"))) == 0) && 
+            tran %in% c("log", "log1p", "log2", "log10", "sqrt", "logit", "probit", 
+                        "exp", "identity"))
+            return(get(paste(tran, "trans", sep = "_"), envir = asNamespace("scales"))())
+        # not built-in, so let's get a list
+        tran = .make.link(tran)
+    }
+    # tran is a list. we'll incorporate any scaling
+    tran$mult = ifelse(is.null(misc$tran.mult), 1, misc$tran.mult)
+    tran$offset = ifelse(is.null(misc$tran.offset), 0, misc$tran.offset)
+    with(tran, 
+         scales::trans_new(name, 
+                           function(x) mult * linkfun(x + offset), 
+                           function(z) linkinv(z / mult) - offset))
+}
+