+++ R/emmGrid-methods.R
@@ -662,13 +662,13 @@ emm_defaults = list (
 #' @export
 #' @param x an \code{emmGrid} object
 #' @param value \code{list} or replacement levels. See the documentation for
-#' \code{\link{update.emmGrid}} with the \code{levels} argument, 
+#' \code{update.emmGrid} with the \code{levels} argument, 
 #' as well as the section below on \dQuote{Replaciong levels}
 #'   
 #' @return \code{levels<-} replaces the levels of the object in-place.
-#'   See the section on  for details.
+#'   See the section on replacing levels for details.
 #' @section Replacing levels:
-#' The \code{levels<-} method uses \code{\link{update.emmGrid}} to replace the
+#' The \code{levels<-} method uses \code{update.emmGrid} to replace the
 #' levels of one or more factors. This method allows selectively replacing
 #' the levels of just one factor (via subsetting operators), whereas 
 #' \code{update(x, levels = list(...))} requires a list of \emph{all} factors