+++ R/rbind.R
@@ -99,7 +99,7 @@ rbind.emmGrid = function(..., deparse.level = 1, adjust = "bonferroni") {
     obj@levels = lapply(gnms, function(nm) unique(grid[[nm]]))
     names(obj@levels) = gnms
     obj@roles$predictors = setdiff(names(obj@levels), obj@roles$multresp)
-    obj@misc$con.coef = obj@misc$orig.grid = orb@misc$.pairby = NULL
+    obj@misc$con.coef = obj@misc$orig.grid = obj@misc$.pairby = NULL
     update(obj, pri.vars = gnms, by.vars = NULL, adjust = adjust,
            estType = "rbind",
            famSize = n,   # instead of round((1 + sqrt(1 + 8*n)) / 2, 3),
@@ -157,7 +157,7 @@ rbind.emmGrid = function(..., deparse.level = 1, adjust = "bonferroni") {
 #' @examples
 #' 
 #' # After-the-fact 'at' specification
-#' subset(warp.rg, wool == "A")
+#' subset(warp.rg, wool == "A")  ## or warp.rg |> subset(wool == "A")
 #' 
 subset.emmGrid = function(x, subset, ...) {
     sel = eval(substitute(subset), envir = x@grid)