+++ NEWS.md
@@ -18,7 +18,7 @@ emmeans 1.6.0.9xxxxxxxx
   * Fixed bug whereby Scheffe is ignored when there is only one contrast, even
     though `scheffe.rank` > 1 was specified. (#171)
   * Added a `subset()` method for `emmGrid` objects
-  * Bug fix for `mcmc.list` objects (#278)
+  * Bug fixes for `mcmc` and `mcmc.list` objects (#278, #279)
 
 
 emmeans 1.6.0